+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 # Changelog
 
-### vNEXT
+### v0.9.15
+
+- Add support for `graphql` and `@types/graphql` 14.  <br/>
+  [@caiquecastro](https://github.com/caiquecastro) in [#464](https://github.com/apollographql/subscriptions-transport-ws/pull/464)
 
 ### v0.9.14
 - Allow dynamically specifying/overriding the schema in the object returned from `onOperation` [PR #447](https://github.com/apollographql/subscriptions-transport-ws/pull/447)